releasetools: fix wrong suffix given for system_ext
during vintf compatibility check while building ota, only the directories with vintf data in target files zip is extracted. the startswith() function that is responsible for finding and setting right prefix for directories does not add system_ext because "/system" is found as a match before "/system_ext" while searching in the set of known paths, for obvious reasons. moving "/system_ext" to top in the list avoids this error

Change-Id: I22ad15fffd8939a4e773872a7dacd9d4b6ffaeee